Refs #18687. Remove calculate_serialized_size from the type

+++ src/main/java/com/eprosima/fastdds/fastddsgen.java
@@ -659,6 +659,7 @@ public class fastddsgen
             tmanager.addGroup("TypesSource", extensions);
 
             // Load Types common templates
+            tmanager.addGroup("TypesCdrAuxHeader");
             tmanager.addGroup("DDSPubSubTypeHeader");
             tmanager.addGroup("DDSPubSubTypeSource");
 
@@ -789,7 +790,6 @@ public class fastddsgen
                             Utils.writeFile(fileNameH, maintemplates.getTemplate("SerializationHeader"), m_replace);
                 }
 
-                // TODO: Uncomment following lines and create templates
                 if (ctx.existsLastStructure())
                 {
                     m_atLeastOneStructure = true;
@@ -797,21 +797,27 @@ public class fastddsgen
 
                     System.out.println("Generating TopicDataTypes files...");
                     if (returnedValue =
-                            Utils.writeFile(m_outputDir + ctx.getFilename() + "PubSubTypes.h",
-                            maintemplates.getTemplate("DDSPubSubTypeHeader"), m_replace))
+                            Utils.writeFile(m_outputDir + ctx.getFilename() + "CdrAux.hpp",
+                            maintemplates.getTemplate("TypesCdrAuxHeader"), m_replace))
                     {
                         if (returnedValue =
-                                Utils.writeFile(m_outputDir + ctx.getFilename() + "PubSubTypes.cxx",
-                                maintemplates.getTemplate("DDSPubSubTypeSource"), m_replace))
+                                Utils.writeFile(m_outputDir + ctx.getFilename() + "PubSubTypes.h",
+                                    maintemplates.getTemplate("DDSPubSubTypeHeader"), m_replace))
                         {
-                            project.addProjectIncludeFile(ctx.getFilename() + "PubSubTypes.h");
-                            project.addProjectSrcFile(ctx.getFilename() + "PubSubTypes.cxx");
-                            if (m_python)
+                            if (returnedValue =
+                                    Utils.writeFile(m_outputDir + ctx.getFilename() + "PubSubTypes.cxx",
+                                        maintemplates.getTemplate("DDSPubSubTypeSource"), m_replace))
                             {
-                                System.out.println("Generating Swig interface files...");
-                                returnedValue = Utils.writeFile(
-                                    m_outputDir + ctx.getFilename() + "PubSubTypes.i",
-                                    maintemplates.getTemplate("DDSPubSubTypeSwigInterface"), m_replace);
+                                project.addProjectIncludeFile(ctx.getFilename() + "CdrAux.hpp");
+                                project.addProjectIncludeFile(ctx.getFilename() + "PubSubTypes.h");
+                                project.addProjectSrcFile(ctx.getFilename() + "PubSubTypes.cxx");
+                                if (m_python)
+                                {
+                                    System.out.println("Generating Swig interface files...");
+                                    returnedValue = Utils.writeFile(
+                                            m_outputDir + ctx.getFilename() + "PubSubTypes.i",
+                                            maintemplates.getTemplate("DDSPubSubTypeSwigInterface"), m_replace);
+                                }
                             }
                         }
                     }
